+++ odoo/local-src/specific_stock/tests/test_stock_inventory.py
@@ -44,9 +44,8 @@ class TestStockInventory(TransactionCase):
         # the same than the create date of the line.
         # It's why I check that the date_last_inventory is greater or equal
         # to the create_date of the line.
-        self.assertTrue(
-            product.date_last_inventory >= inventory_line.create_date
-        )
+        self.assertGreaterEqual(product.date_last_inventory,
+                                inventory_line.create_date)
         current_date_last_inventory = product.date_last_inventory
 
         # Now we wil update the quantity on hand with the wizard
@@ -62,6 +61,36 @@ class TestStockInventory(TransactionCase):
 
         self.assertEqual(product.date_last_inventory,
                          current_date_last_inventory)
+
+    @post_install(True)
+    @at_install(False)
+    def test_create_daily_inventory(self):
+        """
+        Test if the method test_create_daily_inventory skip all weekends days
+        and bank holidays.
+        :return:
+        """
+        config_param = self.env['ir.config_parameter']
+        config_param.set_param('stock.price_limit_for_inventory', 100)
+        config_param.set_param('stock.nbr_open_days', 220)
+
+        date_test = date(2017, 1, 31)
+        self.env['bank.holiday'].create({
+            'name': 'Test',
+            'date': fields.Date.to_string(date_test)
+        })
+
+        inventory = self.env['stock.inventory'].create_daily_inventory(
+            date_today_overwrite=date_test
+        )
+        self.assertIsNone(inventory)
+
+        # 1 january is a Sunday
+        date_test = date(2017, 1, 1)
+        inventory = self.env['stock.inventory'].create_daily_inventory(
+            date_today_overwrite=date_test
+        )
+        self.assertIsNone(inventory)
 
     @post_install(True)
     @at_install(False)
@@ -122,7 +151,17 @@ class TestStockInventory(TransactionCase):
         - 10 products with sales orders
         :return:
         """
-        product = self.env['product.product']
+        # Change the fiscal year to have a year from 1 january to 31 december
+        company = self.env.user.company_id
+        company.write({
+            'fiscalyear_last_month': 12,
+            'fiscalyear_last_day': 31,
+        })
+        today_year = date.today().year
+
+        product_obj = self.env['product.product']
+        so_obj = self.env['sale.order']
+        stock_obj = self.env['stock.inventory']
         uom_id = self.ref('product.product_uom_unit')
 
         partner = self.env['res.partner'].create({
@@ -136,372 +175,120 @@ class TestStockInventory(TransactionCase):
         ir_config.set_param('stock.delay_inventory_best_sellers_products', 6)
         ir_config.set_param('stock.delay_inventory_other_products', 12)
         ir_config.set_param('stock.months_between_inventory', 2)
+        ir_config.set_param('stock.best_sellers_duration', 12)
 
-        # I need to set a date_last_inventory to all products
-        # to ignore existing products during tests
+        # Disable all products to take only new products in count
         update_date_last_inventory_query = """
-        UPDATE product_product SET date_last_inventory = NOW();
+        UPDATE product_product SET active = FALSE;
         """
         self.env.cr.execute(update_date_last_inventory_query)
 
         # Create products (12 products)
         #######################
-        # The best seller 1 has 2 sales orders
-        # The best seller 2 has 3 sales orders
-        # Two products with price greater than 5000
-        product_1_sec_best_seller = product.create({
-            'name': 'Product 1 (seconde best seller)',
-            'uom_id': uom_id,
-            'list_price': 3000
-        })
-        product_2_best_seller = product.create({
-            'name': 'Product 2 (best seller)',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_3 = product.create({
-            'name': 'Product 3',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_4_sec_expensive = product.create({
-            'name': 'Product 4 (second expensive product)',
-            'uom_id': uom_id,
-            'list_price': 10000
-        })
-        product_5 = product.create({
-            'name': 'Product 5',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_6 = product.create({
-            'name': 'Product 6',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_7_most_expensive = product.create({
-            'name': 'Product 7 (most expensive)',
-            'uom_id': uom_id,
-            'list_price': 15000
-        })
-        product_8 = product.create({
-            'name': 'Product 8',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_9 = product.create({
-            'name': 'Product 9',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_10 = product.create({
-            'name': 'Product 10',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_11 = product.create({
-            'name': 'Product 11',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
-        product_12 = product.create({
-            'name': 'Product 12',
-            'uom_id': uom_id,
-            'list_price': 1000
-        })
+        # - The product 1 has two SO
+        # - The product 2 has three SO
+        # - The product 4 and 8 are an expensive product (> 5K)
+        default_price = 1000
+        product_prices = {
+            'product_4': 10000,
+            'product_7': 15000,
+        }
 
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 1',
-                    'product_id': product_1_sec_best_seller.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 1',
-                    'product_id': product_1_sec_best_seller.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 2',
-                    'product_id': product_2_best_seller.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 2',
-                    'product_id': product_2_best_seller.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 2',
-                    'product_id': product_2_best_seller.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 3',
-                    'product_id': product_3.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 4',
-                    'product_id': product_4_sec_expensive.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 5',
-                    'product_id': product_5.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 6',
-                    'product_id': product_6.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 7',
-                    'product_id': product_7_most_expensive.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 8',
-                    'product_id': product_8.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 9',
-                    'product_id': product_9.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 10',
-                    'product_id': product_10.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 11',
-                    'product_id': product_11.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
-        self.env['sale.order'].create({
-            'partner_id': partner.id,
-            'order_line': [
-                (0, 0, {
-                    'name': 'Product 12',
-                    'product_id': product_12.id,
-                    'product_uom': uom_id,
-                    'product_uom_qty': 10,
-                    'sequence': 1,
-                }),
-            ]
-        })
+        default_nbr_of_so = 1
+        nbr_of_so = {
+            'product_1': 2,
+            'product_2': 3,
+        }
+
+        for i in range(1, 13):
+            # Create the product
+            product_name = 'product_%s' % i
+            product = product_obj.create({
+                'name': product_name,
+                'uom_id': uom_id,
+                'list_price': product_prices.get(product_name, default_price),
+            })
+            setattr(self, product_name, product)
+
+            # Create sale orders
+            number_of_so = nbr_of_so.get(product_name, default_nbr_of_so)
+            for x in range(number_of_so):
+                so_obj.create({
+                    'partner_id': partner.id,
+                    'order_line': [
+                        (0, 0, {
+                            'name': 'Line for SO',
+                            'product_id': product.id,
+                            'product_uom': uom_id,
+                            'product_uom_qty': 10,
+                            'sequence': 1,
+                        })
+                    ]
+                })
 
         # Create a first inventory
-        date_today_overwrite = date(year=2017, month=1, day=1)
-        periods = {
-            'expensive': {
-                'date_start': '2017-01-01',
-                'date_end': '2017-01-01',
-                'delay': 6,
-                'nbr_inventory_per_year': 2,
-            },
-            'best_sellers': {
-                'date_start': '2017-01-01',
-                'date_end': '2017-01-01',
-                'delay': 6,
-                'nbr_inventory_per_year': 2,
-            },
-            'other': {
-                'date_start': '2017-01-01',
-                'date_end': '2017-01-02',
-                'delay': 12,
-                'nbr_inventory_per_year': 1,
-            },
-        }
-        inventory_products = product.get_products_daily_inventory(
-            periods, date_today_overwrite=date_today_overwrite
+        date_today_overwrite = date(year=today_year, month=6, day=1)
+        self.env['bank.holiday'].search(
+            [('date', '=', fields.Date.to_string(date_today_overwrite))]
+        ).unlink()
+
+        inventory = stock_obj.create_daily_inventory(
+            date_today_overwrite=date_today_overwrite,
         )
+        inventory_products = inventory.line_ids.mapped('product_id')
 
         # I must have 8 products in this inventory
         # - 2 expensive product
         #       nbr of expensive products / (days / nbr inv per year)
         #       2 / (2 / 2) = 1 product
         # - 2 best seller
-        #       20% of nbr of products with so / (days / nbr inv per year)
-        #       20% of 10 / (2 / 2) = 2 products
+        #       nbr of best sellers / (days / nbr inv per year)
+        #       2 / (2 / 2) = 2 products
         # - 4 other products
         #       nbr of products / (days / nbr inv per year)
         #       8 products left / (2 / 1) = 4 products
         self.assertEqual(len(inventory_products), 8)
 
-        # Check if one of the two most expensive product is in the inventory
-        self.assertTrue(product_7_most_expensive in inventory_products
-                        and product_4_sec_expensive in inventory_products)
+        # Check if all expensive products are in the inventory
+        self.assertTrue(self.product_7 in inventory_products
+                        and self.product_4 in inventory_products)
 
-        # One of the two best sellers should be in the inventory
-        self.assertTrue(product_2_best_seller in inventory_products
-                        and product_1_sec_best_seller in inventory_products)
-
-        inventory = self.env['stock.inventory'].create({
-            'name': 'Daily inventory: %s' % fields.Date.today(),
-            'filter': 'partial',
-        })
-
-        location_id = self.env.ref('stock.stock_location_stock').id
-        for product in inventory_products:
-            inventory.line_ids.create({
-                'inventory_id': inventory.id,
-                'product_id': product.id,
-                'product_qty': 20,
-                'location_id': location_id
-            })
+        # Check if all best sellers products are in the inventory
+        self.assertTrue(self.product_1 in inventory_products
+                        and self.product_2 in inventory_products)
 
         # Validate the inventory
         inventory.action_done()
-        # The method action_done will write the date_last_inventory with today
-        # or the real date_last_inventory should be 2017-01-01
-        inventory_products.write({
-            'date_last_inventory': '2017-01-01'
-        })
 
-        # Create a first inventory
-        date_today_overwrite = date(year=2017, month=1, day=2)
-        new_periods = {
-            'expensive': {
-                'date_start': '2017-01-02',
-                'date_end': '2017-01-02',
-                'delay': 6,
-                'nbr_inventory_per_year': 2,
-            },
-            'best_sellers': {
-                'date_start': '2017-01-02',
-                'date_end': '2017-01-02',
-                'delay': 6,
-                'nbr_inventory_per_year': 2,
-            },
-            'other': {
-                'date_start': '2017-01-01',
-                'date_end': '2017-01-02',
-                'delay': 12,
-                'nbr_inventory_per_year': 1,
-            },
-        }
-        new_inventory_products = product.get_products_daily_inventory(
-            new_periods, date_today_overwrite=date_today_overwrite
+        # Create a inventory the next day month
+        date_today_overwrite = date(year=today_year, month=7, day=20)
+        self.env['bank.holiday'].search(
+            [('date', '=', fields.Date.to_string(date_today_overwrite))]
+        ).unlink()
+
+        inventory = stock_obj.create_daily_inventory(
+            date_today_overwrite=date_today_overwrite,
         )
-        # I must have 4 products in this inventory
-        # - 0 expensive product
-        # - 0 best seller
+        inventory_products = inventory.mapped('line_ids.product_id')
+
+        # This inventory will contains 4 products
+        # - 0 expensive products
+        # - 0 best sellers products
         # - 4 other products
-        #       nbr of products / (days / nbr inv per year)
-        #       8 products left / (2 / 1) = 4 products
-        self.assertEqual(len(new_inventory_products), 4)
+        # We have no expensive products and best sellers because
+        # we want two months between two inventory.
+        self.assertEqual(len(inventory_products), 4)
 
         # We'll not validate this inventory and create a new one
-        ir_config.set_param('stock.months_between_inventory', '0')
+        # I'll change the duration between two inventory
+        ir_config.set_param('stock.months_between_inventory', '1')
 
-        new_inventory_products = product.get_products_daily_inventory(
-            new_periods, date_today_overwrite=date_today_overwrite
+        inventory = stock_obj.create_daily_inventory(
+            date_today_overwrite=date_today_overwrite,
         )
-        # I must have 4 products in this inventory
+        inventory_products = inventory.mapped('line_ids.product_id')
+        # This inventory will contains 8 products
         # - 2 expensive product
         # - 2 best seller
         # - 4 other products
-        #       nbr of products / (days / nbr inv per year)
-        #       8 products left / (2 / 1) = 4 products
-        self.assertEqual(len(new_inventory_products), 8)
+        self.assertEqual(len(inventory_products), 8)